feat: error on missing target (instead of autoSrc)
Previously, the default value of `defaultRule`, which is used when a
target name doesn't match any rule, was to just treat it as a source
file. However, this meant that if it was an accident, users would only
see a tiny error line during build complaining about `cp` not being able
to find a file.

This commit instead changes the default behavior to throwing an error
about a missing target. This is a Nix `throw`, and therefore it will be
surfaced while evaluating the makefile, instead of during build. And
since I was in such a good mood after implementing patterns (cf next
commits), I even wrote a nice little error message:

    error: need to make target 'hello.c', but no rule was found for it.

    hint: in case 'hello.c' is a source file, you should add a rule
          for it:
              rules = {
                  "hello.c" = nix-make.utils.autoSrc;
              };

    note: you might want to use pattern rules (e.g. '%.c') to specify
          source files:
              rules = {
                  "%.c" = nix-make.utils.autoSrc;
              };

    note: alternatively, you can also set the `defaultRule` top-level
          attribute if you want to change the default behavior when a
          target isn't found:
              defaultRule = targetName: nix-make.utils.autoSrc;