fix: handle `make`/`dep` calls on derivations
this has the nice side-effect of allowing calls to dep to be chained,
e.g. `dep (dep "hello.o")` now works. this could happen when if a
recipe-factory was called with a "filename" that was actually the result
of a call to `dep`, and then the recipe called `dep` on it again.

for example:

```nix
let
  cc = file: { dep, ... }: run ''
    gcc -c ${dep file} -o $out
  '';
in {
  rules = {
    "hello.o" = { dep, ... }: cc (dep "hello.c");
  }
}
```